# Perf: improve EventProcessor#processEvent(E) performance