--- a/resilience4j-all/src/main/java/io/github/resilience4j/decorators/Decorators.java
+++ b/resilience4j-all/src/main/java/io/github/resilience4j/decorators/Decorators.java
@@ -27,7 +27,7 @@ import java.util.function.*;
 /**
  * A Decorator builder which can be used to apply multiple decorators to a Supplier, Callable
  * Function, Runnable, CompletionStage or Consumer.
- * <p></p>
+ * <p>
  * Decorators are applied in the order of the builder chain. For example, consider:
  *
  * <pre>{@code