--- a/resilience4j-circuitbreaker/src/jmh/java/io/github/resilience4j/circuitbreaker/CircuitBreakerBenchmark.java
+++ b/resilience4j-circuitbreaker/src/jmh/java/io/github/resilience4j/circuitbreaker/CircuitBreakerBenchmark.java
@@ -35,12 +35,13 @@ import java.util.function.Supplier;
 public class CircuitBreakerBenchmark {
 
     private static final int ITERATION_COUNT = 10;
-    private static final int WARMUP_COUNT = 10;
+    private static final int WARMUP_COUNT = 3;
     private static final int THREAD_COUNT = 2;
     private static final int FORK_COUNT = 2;
 
     private Supplier<String> protectedSupplier;
-    private Supplier<String> protectedSupplierWithSb;
+    private Supplier<String> protectedSupplierWithOneConsumer;
+    private Supplier<String> protectedSupplierWithDiffConsumer;
     private Supplier<String> stringSupplier;
 
     public static void main(String[] args) throws RunnerException {
@@ -61,11 +62,17 @@ public class CircuitBreakerBenchmark {
         CircuitBreaker circuitBreaker = CircuitBreaker.ofDefaults("testCircuitBreaker");
         protectedSupplier = circuitBreaker.decorateSupplier(stringSupplier);
 
-        CircuitBreaker circuitBreakerWithSubscriber = CircuitBreaker.ofDefaults("testCircuitBreakerWithSb");
-        circuitBreakerWithSubscriber.getEventPublisher().onEvent(event -> {
-        });
-        protectedSupplierWithSb = CircuitBreaker.decorateSupplier(circuitBreakerWithSubscriber, stringSupplier);
+        CircuitBreaker withOneConsumer = CircuitBreaker.ofDefaults("testCircuitBreakerWithSb");
+        withOneConsumer.getEventPublisher().onEvent(event -> {});
+        protectedSupplierWithOneConsumer = CircuitBreaker.decorateSupplier(withOneConsumer, stringSupplier);
 
+        CircuitBreaker withDiffConsumer = CircuitBreaker.ofDefaults("testCircuitBreakerWithDiffSb");
+        withDiffConsumer.getEventPublisher()
+            .onIgnoredError(event ->{})
+            .onCallNotPermitted(event ->{})
+            .onSuccess(event -> {})
+            .onError(event -> {});
+        protectedSupplierWithDiffConsumer = CircuitBreaker.decorateSupplier(withDiffConsumer, stringSupplier);
     }
 
     @Benchmark
@@ -91,7 +98,16 @@ public class CircuitBreakerBenchmark {
     @Threads(value = THREAD_COUNT)
     @Warmup(iterations = WARMUP_COUNT)
     @Measurement(iterations = ITERATION_COUNT)
-    public String protectedSupplierWithSubscriber() {
-        return protectedSupplierWithSb.get();
+    public String protectedSupplierWithOneConsumer() {
+        return protectedSupplierWithOneConsumer.get();
+    }
+
+    @Benchmark
+    @Fork(value = FORK_COUNT)
+    @Threads(value = THREAD_COUNT)
+    @Warmup(iterations = WARMUP_COUNT)
+    @Measurement(iterations = ITERATION_COUNT)
+    public String protectedSupplierWithDiffConsumer() {
+        return protectedSupplierWithDiffConsumer.get();
     }
 }