… resilience4j-circuitbreaker/src/jmh/resources/resultsWithGCProfiling.txt