Hi, I did some minor performance optimizations for `EventProcessor#processEvent(E)`:
* Use `Class#getName()` instead of `Class#getSimpleName()` to get `List<EventConsumer>`.
* Use `fori` to traverse `List<EventConsumer>` instead of `foreach`.

The following is a comparison of benchmark results before and after optimization:
* Before optimization:
```
Benchmark                                                                                    Mode  Cnt     Score    Error   Units
CircuitBreakerBenchmark.directSupplier                                                      thrpt   20    12.254 ±  0.022  ops/us
CircuitBreakerBenchmark.directSupplier:·gc.alloc.rate                                       thrpt   20    ≈ 10⁻⁴           MB/sec
CircuitBreakerBenchmark.directSupplier:·gc.alloc.rate.norm                                  thrpt   20    ≈ 10⁻⁵             B/op
CircuitBreakerBenchmark.directSupplier:·gc.count                                            thrpt   20       ≈ 0           counts
CircuitBreakerBenchmark.protectedSupplier                                                   thrpt   20     4.018 ±  0.029  ops/us
CircuitBreakerBenchmark.protectedSupplier:·gc.alloc.rate                                    thrpt   20   904.160 ± 75.567  MB/sec
CircuitBreakerBenchmark.protectedSupplier:·gc.alloc.rate.norm                               thrpt   20   248.000 ± 21.382    B/op
CircuitBreakerBenchmark.protectedSupplier:·gc.churn.PS_Eden_Space                           thrpt   20   905.489 ± 75.432  MB/sec
CircuitBreakerBenchmark.protectedSupplier:·gc.churn.PS_Eden_Space.norm                      thrpt   20   248.366 ± 21.358    B/op
CircuitBreakerBenchmark.protectedSupplier:·gc.churn.PS_Survivor_Space                       thrpt   20     0.155 ±  0.016  MB/sec
CircuitBreakerBenchmark.protectedSupplier:·gc.churn.PS_Survivor_Space.norm                  thrpt   20     0.043 ±  0.005    B/op
CircuitBreakerBenchmark.protectedSupplier:·gc.count                                         thrpt   20  2880.000           counts
CircuitBreakerBenchmark.protectedSupplier:·gc.time                                          thrpt   20  1849.000               ms
CircuitBreakerBenchmark.protectedSupplierWithDiffConsumer                                   thrpt   20     3.100 ±  0.076  ops/us
CircuitBreakerBenchmark.protectedSupplierWithDiffConsumer:·gc.alloc.rate                    thrpt   20   969.956 ± 83.507  MB/sec
CircuitBreakerBenchmark.protectedSupplierWithDiffConsumer:·gc.alloc.rate.norm               thrpt   20   344.000 ± 21.382    B/op
CircuitBreakerBenchmark.protectedSupplierWithDiffConsumer:·gc.churn.PS_Eden_Space           thrpt   20   971.674 ± 83.789  MB/sec
CircuitBreakerBenchmark.protectedSupplierWithDiffConsumer:·gc.churn.PS_Eden_Space.norm      thrpt   20   344.608 ± 21.465    B/op
CircuitBreakerBenchmark.protectedSupplierWithDiffConsumer:·gc.churn.PS_Survivor_Space       thrpt   20     0.176 ±  0.021  MB/sec
CircuitBreakerBenchmark.protectedSupplierWithDiffConsumer:·gc.churn.PS_Survivor_Space.norm  thrpt   20     0.062 ±  0.007    B/op
CircuitBreakerBenchmark.protectedSupplierWithDiffConsumer:·gc.count                         thrpt   20  3025.000           counts
CircuitBreakerBenchmark.protectedSupplierWithDiffConsumer:·gc.time                          thrpt   20  1778.000               ms
CircuitBreakerBenchmark.protectedSupplierWithOneConsumer                                    thrpt   20     3.981 ±  0.044  ops/us
CircuitBreakerBenchmark.protectedSupplierWithOneConsumer:·gc.alloc.rate                     thrpt   20   983.143 ± 10.839  MB/sec
CircuitBreakerBenchmark.protectedSupplierWithOneConsumer:·gc.alloc.rate.norm                thrpt   20   272.000 ±  0.001    B/op
CircuitBreakerBenchmark.protectedSupplierWithOneConsumer:·gc.churn.PS_Eden_Space            thrpt   20   984.850 ± 10.901  MB/sec
CircuitBreakerBenchmark.protectedSupplierWithOneConsumer:·gc.churn.PS_Eden_Space.norm       thrpt   20   272.473 ±  0.611    B/op
CircuitBreakerBenchmark.protectedSupplierWithOneConsumer:·gc.churn.PS_Survivor_Space        thrpt   20     0.159 ±  0.014  MB/sec
CircuitBreakerBenchmark.protectedSupplierWithOneConsumer:·gc.churn.PS_Survivor_Space.norm   thrpt   20     0.044 ±  0.004    B/op
CircuitBreakerBenchmark.protectedSupplierWithOneConsumer:·gc.count                          thrpt   20  2885.000           counts
CircuitBreakerBenchmark.protectedSupplierWithOneConsumer:·gc.time                           thrpt   20  1850.000               ms
```

* After optimization:
```
Benchmark                                                                                    Mode  Cnt     Score    Error   Units
CircuitBreakerBenchmark.directSupplier                                                      thrpt   20    12.258 ±  0.019  ops/us
CircuitBreakerBenchmark.directSupplier:·gc.alloc.rate                                       thrpt   20    ≈ 10⁻⁴           MB/sec
CircuitBreakerBenchmark.directSupplier:·gc.alloc.rate.norm                                  thrpt   20    ≈ 10⁻⁵             B/op
CircuitBreakerBenchmark.directSupplier:·gc.count                                            thrpt   20       ≈ 0           counts
CircuitBreakerBenchmark.protectedSupplier                                                   thrpt   20     4.122 ±  0.033  ops/us
CircuitBreakerBenchmark.protectedSupplier:·gc.alloc.rate                                    thrpt   20  1017.650 ±  8.294  MB/sec
CircuitBreakerBenchmark.protectedSupplier:·gc.alloc.rate.norm                               thrpt   20   272.000 ±  0.001    B/op
CircuitBreakerBenchmark.protectedSupplier:·gc.churn.PS_Eden_Space                           thrpt   20  1019.431 ±  8.512  MB/sec
CircuitBreakerBenchmark.protectedSupplier:·gc.churn.PS_Eden_Space.norm                      thrpt   20   272.476 ±  0.609    B/op
CircuitBreakerBenchmark.protectedSupplier:·gc.churn.PS_Survivor_Space                       thrpt   20     0.162 ±  0.015  MB/sec
CircuitBreakerBenchmark.protectedSupplier:·gc.churn.PS_Survivor_Space.norm                  thrpt   20     0.043 ±  0.004    B/op
CircuitBreakerBenchmark.protectedSupplier:·gc.count                                         thrpt   20  2871.000           counts
CircuitBreakerBenchmark.protectedSupplier:·gc.time                                          thrpt   20  1849.000               ms
CircuitBreakerBenchmark.protectedSupplierWithDiffConsumer                                   thrpt   20     3.970 ±  0.023  ops/us
CircuitBreakerBenchmark.protectedSupplierWithDiffConsumer:·gc.alloc.rate                    thrpt   20   894.224 ± 79.317  MB/sec
CircuitBreakerBenchmark.protectedSupplierWithDiffConsumer:·gc.alloc.rate.norm               thrpt   20   248.000 ± 21.382    B/op
CircuitBreakerBenchmark.protectedSupplierWithDiffConsumer:·gc.churn.PS_Eden_Space           thrpt   20   896.456 ± 79.367  MB/sec
CircuitBreakerBenchmark.protectedSupplierWithDiffConsumer:·gc.churn.PS_Eden_Space.norm      thrpt   20   248.621 ± 21.406    B/op
CircuitBreakerBenchmark.protectedSupplierWithDiffConsumer:·gc.churn.PS_Survivor_Space       thrpt   20     0.159 ±  0.014  MB/sec
CircuitBreakerBenchmark.protectedSupplierWithDiffConsumer:·gc.churn.PS_Survivor_Space.norm  thrpt   20     0.044 ±  0.004    B/op
CircuitBreakerBenchmark.protectedSupplierWithDiffConsumer:·gc.count                         thrpt   20  2912.000           counts
CircuitBreakerBenchmark.protectedSupplierWithDiffConsumer:·gc.time                          thrpt   20  1821.000               ms
CircuitBreakerBenchmark.protectedSupplierWithOneConsumer                                    thrpt   20     4.021 ±  0.036  ops/us
CircuitBreakerBenchmark.protectedSupplierWithOneConsumer:·gc.alloc.rate                     thrpt   20  1037.001 ± 47.516  MB/sec
CircuitBreakerBenchmark.protectedSupplierWithOneConsumer:·gc.alloc.rate.norm                thrpt   20   284.000 ± 10.691    B/op
CircuitBreakerBenchmark.protectedSupplierWithOneConsumer:·gc.churn.PS_Eden_Space            thrpt   20  1038.596 ± 47.251  MB/sec
CircuitBreakerBenchmark.protectedSupplierWithOneConsumer:·gc.churn.PS_Eden_Space.norm       thrpt   20   284.439 ± 10.642    B/op
CircuitBreakerBenchmark.protectedSupplierWithOneConsumer:·gc.churn.PS_Survivor_Space        thrpt   20     0.168 ±  0.012  MB/sec
CircuitBreakerBenchmark.protectedSupplierWithOneConsumer:·gc.churn.PS_Survivor_Space.norm   thrpt   20     0.046 ±  0.003    B/op
CircuitBreakerBenchmark.protectedSupplierWithOneConsumer:·gc.count                          thrpt   20  2936.000           counts
CircuitBreakerBenchmark.protectedSupplierWithOneConsumer:·gc.time                           thrpt   20  1840.000               ms
```